requirements.txt
playwright
openai
pydantic
aiohttp
pandas
tqdm
python-dotenv
fastapi
uvicorn
tiktoken
tenacity
